+++ src/Providers.tsx
@@ -57,9 +57,9 @@ export const Providers = withProviders(
   BalancesProvider,
   StakingProvider,
   PoolsConfigProvider,
-  BondedPoolsProvider,
   PoolMembershipsProvider,
   ActivePoolProvider,
+  BondedPoolsProvider,
   ValidatorsProvider,
   UIProvider,
   SubscanProvider,
