+++ src/contexts/Pools/BondedPools/index.tsx
@@ -1,22 +1,20 @@
 // Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import BN from 'bn.js';
 import React, { useState, useEffect, useRef } from 'react';
-import { bnToU8a, u8aConcat } from '@polkadot/util';
 import {
   BondedPool,
   BondedPoolsContextState,
   MaybePool,
   NominationStatuses,
 } from 'contexts/Pools/types';
-import { EMPTY_H256, MOD_PREFIX, U32_OPTS } from 'consts';
 import { AnyApi, AnyMetaBatch, Fn, MaybeAccount } from 'types';
 import { setStateWithRef } from 'Utils';
 import { useStaking } from 'contexts/Staking';
 import { useApi } from '../../Api';
 import { usePoolsConfig } from '../PoolsConfig';
 import { defaultBondedPoolsContext } from './defaults';
+import { useActivePool } from '../ActivePool';
 
 export const BondedPoolsContext = React.createContext<BondedPoolsContextState>(
   defaultBondedPoolsContext
@@ -29,10 +27,10 @@ export const BondedPoolsProvider = ({
 }: {
   children: React.ReactNode;
 }) => {
-  const { api, network, isReady, consts } = useApi();
-  const { enabled } = usePoolsConfig();
+  const { api, network, isReady } = useApi();
   const { getNominationsStatusFromTargets } = useStaking();
-  const { poolsPalletId } = consts;
+  const { enabled, createAccounts } = usePoolsConfig();
+  const { activeBondedPool } = useActivePool();
 
   // stores the meta data batches for pool lists
   const [poolMetaBatches, setPoolMetaBatch]: AnyMetaBatch = useState({});
@@ -62,7 +60,7 @@ export const BondedPoolsProvider = ({
     return () => {
       unsubscribe();
     };
-  }, [network, isReady, enabled]);
+  }, [network, isReady, enabled, activeBondedPool]);
 
   // after bonded pools have synced, fetch metabatch
   useEffect(() => {
@@ -279,31 +277,6 @@ export const BondedPoolsProvider = ({
     };
   };
 
-  // Helper: generates pool stash and reward accounts. assumes poolsPalletId is synced.
-  const createAccounts = (poolId: number) => {
-    const poolIdBN = new BN(poolId);
-    return {
-      stash: createAccount(poolIdBN, 0),
-      reward: createAccount(poolIdBN, 1),
-    };
-  };
-
-  const createAccount = (poolId: BN, index: number): string => {
-    if (!api) return '';
-    return api.registry
-      .createType(
-        'AccountId32',
-        u8aConcat(
-          MOD_PREFIX,
-          poolsPalletId,
-          new Uint8Array([index]),
-          bnToU8a(poolId, U32_OPTS),
-          EMPTY_H256
-        )
-      )
-      .toString();
-  };
-
   const getBondedPool = (poolId: MaybePool) => {
     const pool = bondedPools.find((p: BondedPool) => p.id === poolId) ?? null;
     return pool;
@@ -313,7 +286,6 @@ export const BondedPoolsProvider = ({
     <BondedPoolsContext.Provider
       value={{
         fetchPoolsMetaBatch,
-        createAccounts,
         getBondedPool,
         getPoolNominationStatus,
         getPoolNominationStatusCode,
