+++ src/contexts/Pools/PoolsConfig/defaults.ts
@@ -2,7 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import BN from 'bn.js';
-import { PoolsConfigContextState, PoolStats } from 'contexts/Pools/types';
+import {
+  PoolAddresses,
+  PoolsConfigContextState,
+  PoolStats,
+} from 'contexts/Pools/types';
 
 export const stats: PoolStats = {
   counterForPoolMembers: new BN(0),
@@ -19,6 +23,12 @@ export const defaultPoolsConfigContext: PoolsConfigContextState = {
   enabled: 0,
   addFavourite: (a) => {},
   removeFavourite: (a) => {},
+  createAccounts: (p) => poolAddresses,
   favourites: [],
   stats,
+};
+
+export const poolAddresses: PoolAddresses = {
+  stash: '',
+  reward: '',
 };
